updated module multiple module include logic

+++ modules/adpd188/drivers/adpd188/adpd188bi.c
@@ -462,8 +462,9 @@ static int adpd188bi_sample_fetch(const struct device *dev, enum sensor_channel 
 	    return -EIO;
 	}
     }
-    data->smoke_data.blue_data = buff[0] | (buff[1] << 16);
-    data->smoke_data.ir_data = buff[2] | (buff[3] << 16);
+
+    data->smoke_data.blue_data = (int32_t)buff[0] | ((int32_t)buff[1] << 16);
+    data->smoke_data.ir_data = (int32_t)buff[2] | ((int32_t)buff[3] << 16);
     return 0;
 }
